app/utils/indian_stocks.py
"""
Utility module for handling Indian stock data
This module provides functions to fetch data for Indian stocks
using alternative sources when yfinance fails
"""

import requests
import pandas as pd
import numpy as np
import json
import os
import time
from datetime import datetime, timedelta
import yfinance as yf

# Cache directory
CACHE_DIR = os.path.join(os.path.dirname(os.path.dirname(__file__)), 'data')
os.makedirs(CACHE_DIR, exist_ok=True)

# NSE base URL
NSE_URL = "https://www.nseindia.com/api"
NSE_HEADERS = {
    'user-agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36',
    'accept-encoding': 'gzip, deflate, br',
    'accept-language': 'en-US,en;q=0.9'
}

# Symbol mapping for NSE
NSE_SYMBOL_MAP = {
    'RELIANCE': 'RELIANCE',
    'TCS': 'TCS',
    'HDFCBANK': 'HDFCBANK',
    'INFY': 'INFY',
    'HINDUNILVR': 'HINDUNILVR',
    'ICICIBANK': 'ICICIBANK',
    'SBIN': 'SBIN',
    'BHARTIARTL': 'BHARTIARTL',
    'KOTAKBANK': 'KOTAKBANK',
    'ITC': 'ITC',
    'LT': 'LT',
    'AXISBANK': 'AXISBANK',
    'ASIANPAINT': 'ASIANPAINT',
    'MARUTI': 'MARUTI',
    'HCLTECH': 'HCLTECH',
    'SUNPHARMA': 'SUNPHARMA',
    'BAJFINANCE': 'BAJFINANCE',
    'TATAMOTORS': 'TATAMOTORS',
    'WIPRO': 'WIPRO',
    'NTPC': 'NTPC'
}

def get_nse_session():
    """Create a session for NSE website"""
    session = requests.Session()
    session.headers.update(NSE_HEADERS)

    # Visit homepage first to get cookies
    try:
        session.get("https://www.nseindia.com/", timeout=10)
        return session
    except Exception as e:
        print(f"Error creating NSE session: {e}")
        return None

def get_stock_quote_nse(symbol):
    """Get stock quote from NSE"""
    # Map symbol if needed
    nse_symbol = NSE_SYMBOL_MAP.get(symbol.upper().replace('.NS', ''), symbol.upper().replace('.NS', ''))

    # Check cache first
    cache_file = os.path.join(CACHE_DIR, f"{nse_symbol}_quote.json")
    if os.path.exists(cache_file):
        file_age = datetime.now() - datetime.fromtimestamp(os.path.getmtime(cache_file))
        if file_age < timedelta(hours=1):  # Cache for 1 hour
            with open(cache_file, 'r') as f:
                try:
                    return json.load(f)
                except:
                    pass  # If there's an error reading the cache, fetch fresh data

    # Create a session
    session = get_nse_session()
    if not session:
        return None

    # Get quote data
    try:
        url = f"{NSE_URL}/quote-equity?symbol={nse_symbol}"
        response = session.get(url, timeout=10)
        if response.status_code == 200:
            data = response.json()

            # Cache the data
            with open(cache_file, 'w') as f:
                json.dump(data, f)

            return data
        else:
            print(f"Error fetching NSE quote: {response.status_code}")
            return None
    except Exception as e:
        print(f"Error fetching NSE quote: {e}")
        return None

def get_stock_historical_nse(symbol, days=365):
    """Get historical data from NSE"""
    # Map symbol if needed
    nse_symbol = NSE_SYMBOL_MAP.get(symbol.upper().replace('.NS', ''), symbol.upper().replace('.NS', ''))

    # Check cache first
    cache_file = os.path.join(CACHE_DIR, f"{nse_symbol}_historical_{days}.json")
    if os.path.exists(cache_file):
        file_age = datetime.now() - datetime.fromtimestamp(os.path.getmtime(cache_file))
        if file_age < timedelta(days=1):  # Cache for 1 day
            with open(cache_file, 'r') as f:
                try:
                    data = json.load(f)
                    return pd.DataFrame(data)
                except:
                    pass  # If there's an error reading the cache, fetch fresh data

    # Create a session
    session = get_nse_session()
    if not session:
        return pd.DataFrame()

    # Get historical data
    try:
        end_date = datetime.now().strftime('%d-%m-%Y')
        start_date = (datetime.now() - timedelta(days=days)).strftime('%d-%m-%Y')
        url = f"{NSE_URL}/historical/cm/equity?symbol={nse_symbol}&from={start_date}&to={end_date}"

        response = session.get(url, timeout=10)
        if response.status_code == 200:
            data = response.json()
            if 'data' in data:
                df = pd.DataFrame(data['data'])

                # Rename columns to match yfinance format
                df.rename(columns={
                    'CH_TIMESTAMP': 'Date',
                    'CH_OPENING_PRICE': 'Open',
                    'CH_TRADE_HIGH_PRICE': 'High',
                    'CH_TRADE_LOW_PRICE': 'Low',
                    'CH_CLOSING_PRICE': 'Close',
                    'CH_TOT_TRADED_VAL': 'Volume'
                }, inplace=True)

                # Convert date format
                df['Date'] = pd.to_datetime(df['Date'])
                df.set_index('Date', inplace=True)

                # Sort by date
                df = df.sort_index()

                # Cache the data
                with open(cache_file, 'w') as f:
                    df.reset_index().to_json(f, orient='records')

                return df
            else:
                print(f"No historical data found for {nse_symbol}")
                return pd.DataFrame()
        else:
            print(f"Error fetching NSE historical data: {response.status_code}")
            return pd.DataFrame()
    except Exception as e:
        print(f"Error fetching NSE historical data: {e}")
        return pd.DataFrame()

def get_stock_info_alternative(symbol):
    """Get stock info using alternative sources when yfinance fails"""
    # Try yfinance first
    try:
        # Remove any existing suffix
        base_symbol = symbol.split('.')[0]

        # Try with NSE suffix
        stock = yf.Ticker(f"{base_symbol}.NS")
        info = stock.info
        if info and len(info) > 5:  # Check if we got meaningful data
            return info

        # Try with BSE suffix
        stock = yf.Ticker(f"{base_symbol}.BO")
        info = stock.info
        if info and len(info) > 5:
            return info
    except Exception as e:
        print(f"yfinance error for {symbol}: {e}")

    # If yfinance fails, try NSE API
    try:
        nse_data = get_stock_quote_nse(symbol)
        if nse_data:
            # Convert NSE data to a format similar to yfinance
            info = {
                'symbol': symbol,
                'shortName': nse_data.get('info', {}).get('companyName', ''),
                'longName': nse_data.get('info', {}).get('companyName', ''),
                'regularMarketPrice': nse_data.get('priceInfo', {}).get('lastPrice', 0),
                'regularMarketDayHigh': nse_data.get('priceInfo', {}).get('intraDayHighLow', {}).get('max', 0),
                'regularMarketDayLow': nse_data.get('priceInfo', {}).get('intraDayHighLow', {}).get('min', 0),
                'regularMarketVolume': nse_data.get('securityWiseDP', {}).get('quantityTraded', 0),
                'regularMarketPreviousClose': nse_data.get('priceInfo', {}).get('previousClose', 0),
                'regularMarketOpen': nse_data.get('priceInfo', {}).get('open', 0),
                'marketCap': nse_data.get('securityInfo', {}).get('issuedSize', 0) * nse_data.get('priceInfo', {}).get('lastPrice', 0),
                'fiftyTwoWeekHigh': nse_data.get('priceInfo', {}).get('weekHighLow', {}).get('max', 0),
                'fiftyTwoWeekLow': nse_data.get('priceInfo', {}).get('weekHighLow', {}).get('min', 0),
                'trailingPE': nse_data.get('metadata', {}).get('pdSectorPe', 0),
                'dividendYield': nse_data.get('securityInfo', {}).get('isinDivPayDate', 0),
                'industry': nse_data.get('metadata', {}).get('industry', ''),
                'sector': nse_data.get('metadata', {}).get('sector', '')
            }
            return info
    except Exception as e:
        print(f"NSE API error for {symbol}: {e}")

    # If all fails, create a minimal info object with dummy data
    return {
        'symbol': symbol,
        'shortName': symbol,
        'longName': symbol,
        'regularMarketPrice': 0,
        'regularMarketDayHigh': 0,
        'regularMarketDayLow': 0,
        'regularMarketVolume': 0,
        'regularMarketPreviousClose': 0,
        'regularMarketOpen': 0,
        'marketCap': 0,
        'fiftyTwoWeekHigh': 0,
        'fiftyTwoWeekLow': 0
    }

def get_historical_data_alternative(symbol, period='1y'):
    """Get historical data using alternative sources when yfinance fails"""
    # Remove any existing suffix
    base_symbol = symbol.split('.')[0]

    # Try with NSE suffix first
    try:
        nse_symbol = f"{base_symbol}.NS"
        stock = yf.Ticker(nse_symbol)
        data = stock.history(period=period)
        if not data.empty:
            print(f"Successfully fetched data for {nse_symbol} using yfinance")
            return data
    except Exception as e:
        print(f"Error fetching data for {base_symbol}.NS: {e}")

    # Try with BSE suffix
    try:
        bse_symbol = f"{base_symbol}.BO"
        stock = yf.Ticker(bse_symbol)
        data = stock.history(period=period)
        if not data.empty:
            print(f"Successfully fetched data for {bse_symbol} using yfinance")
            return data
    except Exception as e:
        print(f"Error fetching data for {base_symbol}.BO: {e}")

    # If yfinance fails, try using sample data for testing
    print(f"Using sample data for {symbol}")

    # Create sample data for testing
    dates = pd.date_range(end=pd.Timestamp.now(), periods=365)
    data = pd.DataFrame({
        'Open': np.random.normal(100, 5, len(dates)),
        'High': np.random.normal(105, 5, len(dates)),
        'Low': np.random.normal(95, 5, len(dates)),
        'Close': np.random.normal(100, 5, len(dates)),
        'Volume': np.random.randint(100000, 1000000, len(dates))
    }, index=dates)

    # Ensure High is always higher than Open, Close, and Low
    for i in range(len(data)):
        data.iloc[i, 1] = max(data.iloc[i, 0], data.iloc[i, 1], data.iloc[i, 3]) + 1

    # Ensure Low is always lower than Open, Close, and High
    for i in range(len(data)):
        data.iloc[i, 2] = min(data.iloc[i, 0], data.iloc[i, 2], data.iloc[i, 3]) - 1

    return data

def get_current_price_alternative(symbol):
    """Get current price using alternative sources when yfinance fails"""
    # Try yfinance first
    try:
        # Remove any existing suffix
        base_symbol = symbol.split('.')[0]

        # Try with NSE suffix
        stock = yf.Ticker(f"{base_symbol}.NS")
        data = stock.history(period='1d')
        if not data.empty:
            return data['Close'].iloc[-1]

        # Try with BSE suffix
        stock = yf.Ticker(f"{base_symbol}.BO")
        data = stock.history(period='1d')
        if not data.empty:
            return data['Close'].iloc[-1]
    except Exception as e:
        print(f"yfinance error for {symbol}: {e}")

    # If yfinance fails, try NSE API
    try:
        nse_data = get_stock_quote_nse(symbol)
        if nse_data and 'priceInfo' in nse_data:
            return nse_data['priceInfo'].get('lastPrice', 0)
    except Exception as e:
        print(f"NSE API error for {symbol}: {e}")

    # If all fails, return 0
    return 0
