requirements.txt
flask==2.3.3
dash==2.14.1
dash-bootstrap-components==1.5.0
plotly==5.18.0
pandas==2.1.1
numpy==1.26.0
yfinance==0.2.31
python-dotenv==1.0.0
flask-sqlalchemy==3.1.1
openai==0.28.1
scikit-learn==1.3.2
gunicorn==21.2.0
flask-login==0.6.3
matplotlib==3.8.0
seaborn==0.13.0
openpyxl==3.1.2
Pillow==10.1.0
